Reworked how apps were referenced in streams and remote activity prompts.
The new format for "apps", when defined in the context of ptreams and remote activity prompts, is to have it be an object whose field names are the platforms and whose field values are the same as before without the "platform" field.

Also, a bug has been fixed for the OhmageNumber references in conditions.

Closes #724
Closes #727